requirements.txt
Django==4.1.4
django-simple-history==3.2.0
djangorestframework-simplejwt==5.2.2